[#65361080] Writes inline images to disk
If an image lacks a fileName, we create a 'dummy' filename.

If the path to the image does not exist, we call osgDB::writeImageFile to create
it, otherwise, we just reference the existing path.
Note that if an image is modified but its fileName remains unchanged then the
modification will *not* be taken into account.